+++ atmega32u4/hardware.c
@@ -12,15 +12,19 @@ void hardware_init(void) {
     // set for 16 MHz clock
     CPU_PRESCALE(0);
 
+    // NOTE: the security mechanism on the MCUCR register requires that we write
+    // The JTD bit twice before changes will take effect.
+    {
+        uint8_t new_value = MCUCR | _BV(JTD);
+        MCUCR = new_value;
+        MCUCR = new_value;
+    }
+
     wdt_reset();
     wdt_disable();
 }
 
 void led_init(void) {
-}
-
-void reset_mcu(void) {
-
 }
 
 #if defined(BOOTLOADER_ATMEL_DFU)
@@ -31,17 +35,25 @@ void bootloader_jmp(void) {
     run_bootloader();
 }
 
+void reset_mcu(void) {
+    wdt_enable(WDTO_15MS);
+    wdt_reset();
+    while (1);
+}
+
+
 #elif defined(BOOTLOADER_KP_BOOT_32U4)
 
 #include "bootloaders/kp_boot_32u4/interface/kp_boot_32u4.h"
 
-void bootloader_jmp(void) {
-    cli();
-    wdt_enable(WDTO_15MS);
-    wdt_reset();
-    *(uint16_t*)MAGIC_ADDRESS = MAGIC_ENTER_BOOT;
-    while (1);
+void reset_mcu(void) {
+    kp_boot_reset();
 }
+
+void bootloader_jmp(void) {
+    kp_boot_jmp();
+}
+
 #endif
 
 void wdt_init(void) {