+++ atmega32u4/main.c
@@ -11,7 +11,8 @@
 #include "usb_reports/media_report.h"
 #include "usb_reports/mouse_report.h"
 #include "usb_reports/vendor_report.h"
-#include "core/timer.h"
+
+#include "core/error.h"
 #include "core/flash.h"
 #include "core/hardware.h"
 #include "core/macro.h"
@@ -19,8 +20,10 @@
 #include "core/matrix_scanner.h"
 #include "core/usb_commands.h"
 #include "core/settings.h"
+#include "core/timer.h"
 
 #include "key_handlers/key_hold.h"
+#include "key_handlers/key_mouse.h"
 
 #include "bootloaders/kp_boot_32u4/interface/kp_boot_32u4.h"
 
@@ -42,6 +45,17 @@ void setup_everything(void) {
     matrix_scanner_init();
     reset_usb_reports();
     keyboards_init();
+}
+
+/// When we encounter a critical error, stop all other features and only
+/// enable those necessary to load a new valid layout.
+void recovery_mode_main_loop(void) {
+    while (1) {
+        // usb out reports
+        send_vendor_report();
+        handle_vendor_out_reports();
+        wdt_kick();
+    }
 }
 
 int main(void) {
@@ -67,6 +81,7 @@ int main(void) {
         // passthrough_keycodes_task();
         interpret_all_keyboard_matrices();
 
+        mouse_key_task();
         macro_task();
 
         send_keyboard_report();
@@ -78,8 +93,13 @@ int main(void) {
         handle_vendor_out_reports();
 
         sticky_key_task();
-        hold_key_task();
+        hold_key_task(false);
+
+        if (has_critical_error()) {
+            recovery_mode_main_loop();
+        }
 
         wdt_kick();
+
     }
 }