+++ src/core/matrix_interpret.c
@@ -32,6 +32,7 @@
 #include "core/util.h"
 
 #include "key_handlers/key_handlers.h"
+#include "key_handlers/key_hold.h"
 
 #include "usb_reports/keyboard_report.h"
 
@@ -52,6 +53,8 @@ XRAM key_event_queue_t s_key_event_queues[2];
 
 XRAM uint8_t s_has_dirty_matrix;
 XRAM uint8_t s_has_dirty_event_queue;
+
+XRAM uint8_t s_buffered_key_len;
 
 bit_t g_input_disabled = false;
 bit_t dongle_active = true;
@@ -84,6 +87,7 @@ static XRAM uint8_t s_sticky_has_stuck_layer;
 
 static void keyboard_trigger_event(keycode_t keycode, key_event_t event) REENT;
 static void keyboard_reset_event_handlers(void);
+static keycode_t get_keycode_from_layer(layer_mask_t layer_mask, uint8_t row, uint8_t col) REENT;
 
 void layer_queue_add(uint8_t layer) {
     s_layer_dirty = 1;
@@ -156,15 +160,16 @@ void layer_queue_apply(uint8_t kb_slot_id) {
     flush_queues();
 }
 
-static bool sticky_relase_timer_done(void) {
+static
+bit_t sticky_relase_timer_done(void) {
     return (uint16_t)(timer_read16_ms() - s_sticky_clear_start_time) > STICKY_KEY_RELEASE_DELAY;
 }
 
-uint8_t get_active_slot_id(void) REENT {
+uint8_t get_active_slot_id(void) {
     return s_active_slot;
 }
 
-uint8_t get_active_keyboard_id(void) REENT {
+uint8_t get_active_keyboard_id(void) {
     return g_keyboard_slots[s_active_slot].kb_id;
 }
 
@@ -215,15 +220,29 @@ static void apply_event_trigger_queue(uint8_t keyboard_id) {
     key_event_queue_t XRAM* queue = &s_key_event_queues[READ_EVENT_QUEUE()];
 
     for (i = 0; i < queue->length; ++i) {
-
+        const uint8_t type = queue->events[i].type;
         if (queue->events[i].keyboard_id != keyboard_id) {
             continue;
         }
 
-        keyboard_trigger_event(
-            queue->events[i].keycode,
-            queue->events[i].type
-        );
+        if (type == EVENT_BUFFERED_KEY_PRESS0 ||
+            type == EVENT_BUFFERED_KEY_PRESS1) {
+            const uint8_t key_num = queue->events[i].keycode;
+            const layer_mask_t active_layer =
+                keyboard_get_layer_mask(get_slot_id(keyboard_id));
+            const keycode_t keycode =
+                get_keycode_from_layer(active_layer, key_num/8, key_num%8);
+            if (type == EVENT_BUFFERED_KEY_PRESS1) {
+                queue_keycode_event(key_num, EVENT_BUFFERED_KEY_PRESS0, keyboard_id);
+            } else if (type == EVENT_BUFFERED_KEY_PRESS0) {
+                keyboard_trigger_event(keycode, EVENT_PRESSED);
+            }
+        } else {
+            keyboard_trigger_event(
+                queue->events[i].keycode,
+                queue->events[i].type
+            );
+        }
     }
 }
 
@@ -298,7 +317,7 @@ uint8_t get_slot_id(uint8_t kb_id) {
     return s_slot_id_map[kb_id];
 }
 
-bool is_keyboard_active(uint8_t kb_id) {
+bit_t is_keyboard_active(uint8_t kb_id) {
     return get_slot_id(kb_id) != INVALID_DEVICE_ID;
 }
 
@@ -368,7 +387,7 @@ void keyboards_init(void) {
 // TODO: update this
 // NOTE: this should not be called from an interrupt, as it will break
 // how the matrix interpreting step.
-void keyboard_update_device_matrix(uint8_t device_id, const uint8_t *matrix_packet) REENT {
+void keyboard_update_device_matrix(uint8_t device_id, const XRAM uint8_t *matrix_packet) REENT {
     // first figure out what sort of packet was received and
     const uint8_t packet_type = matrix_packet[0] >> PACKET_MATRIX_TYPE_BIT_POS;
     const uint8_t packet_data_size = matrix_packet[0] & PACKET_MATRIX_SIZE_MASK;
@@ -378,11 +397,11 @@ void keyboard_update_device_matrix(uint8_t device_id, const uint8_t *matrix_pack
     const uint8_t device_matrix_size = GET_SETTING(layout.devices[device_id].matrix_size);
 
     // matrix_data now points to the start of the key list
-    const uint8_t* matrix_data = &matrix_packet[1];
+    const XRAM uint8_t* matrix_data = &matrix_packet[1];
 
     // get matrix slot from kb_id
 
-    uint8_t *matrix_write_pos;
+    XRAM uint8_t* matrix_write_pos;
 
     uint8_t kb_slot_id = get_slot_id(kb_id);
 
@@ -622,8 +641,8 @@ void keyboard_interpret_matrix(uint8_t kb_slot_id) {
     keyboard->layer_changed = false;
     keyboard->is_dirty = 0;
 
+    s_buffered_key_len = 0;
     s_active_slot = kb_slot_id;
-
 
     active_layer = keyboard_get_layer_mask(kb_slot_id);
     start_layer = get_partial_layer_mask(kb_slot_id);
@@ -661,6 +680,21 @@ void keyboard_interpret_matrix(uint8_t kb_slot_id) {
                 } else if (pressed) {
                     event = EVENT_PRESSED;
                     keyboard->num_keys_down += 1;
+
+                    // TODO: make this a more generic mechanism?
+                    if (hold_key_buffer_other_keys() || (s_buffered_key_len>0)) {
+                        const uint8_t key_num = byte*8 + bit;
+
+                        // If s_buffered_key_len > 0, then that means we have
+                        // already started adding keys to the buffer, and don't
+                        // need to retrigger the hold key task.
+                        if (s_buffered_key_len == 0) {
+                            hold_key_task(true);
+                        }
+                        s_buffered_key_len++;
+                        queue_keycode_event(key_num, EVENT_BUFFERED_KEY_PRESS1, keyboard->kb_id);
+                        continue;
+                    }
                 } else if (released) {
                     event = EVENT_RELEASED;
                     keyboard->num_keys_down -= 1;
@@ -707,14 +741,15 @@ void keyboard_interpret_matrix(uint8_t kb_slot_id) {
         }
     }
 
+    // wait until all layer and key events have been processed before handling
+    // modifiers
+    apply_mods();
+
+
     /* // if we changed layers, notify the host on the custom interface */
     /* if (keyboard->layer_changed) { */
     /*  cmd_send_layer(kb_slot_id); */
     /* } */
-
-    // wait until all layer and key events have been processed before handling
-    // modifiers
-    apply_mods();
 
     memcpy(keyboard->matrix_prev, keyboard->matrix, keyboard->matrix_size);
 }